refactor: move last filters to param builder folder
